thieu payment

+++ Model/Product/CartItem.cs
@@ -8,7 +8,7 @@ namespace ArtsShop.Model.Product
         public int Id { get; set; }
         public int ProductId { get; set; }
         public int CartId { get; set; }
-        public double Price { get; set; }
+        public Decimal Price { get; set; }
         public int Discount { get; set; }
         public int Quantity { get; set; }
         public string Image { get; set; }
